Update training view and project cleaning
Split classes and structs to separate files.
Remove unused code.
Add timer and automatic exercise collapsing in training view.
