+++ GymTracker/View/TrainingView/TrainingView.swift
@@ -16,16 +16,15 @@ struct TrainingView: View {
 	
 	@Namespace var bottomId
 	
-	@State private var viewModel: ViewModel
 	@State private var showingNewExerciseView = false
 	
 	var body: some View {
 		VStack {
-			Text(training.name)
+			TimerView(from: Date(timeIntervalSince1970: training.startDate))
 			
 			Button("End training", role: .destructive) {
 				ExerciseRecordManager.shared.addOrUpdateStats(training.exercises)
-				
+				training.finishTraining()
 				dismiss()
 			}
 			.buttonStyle(.bordered)
@@ -33,7 +32,7 @@ struct TrainingView: View {
 			ScrollViewReader { proxy in
 				ScrollView {
 					ForEach(training.exercises) { exercise in
-						ExerciseView(exercise: exercise, viewModel: viewModel, showTextField: training.exercises.last == exercise) {
+						ExerciseView(exercise: exercise, showTextField: training.exercises.last == exercise) {
 							scrollToBottom(proxy)
 						}
 					}
@@ -54,12 +53,12 @@ struct TrainingView: View {
 			NewExerciseView { newExercise in
 				training.addExercise(newExercise)
 			}
+			.presentationDetents([.medium, .large])
 		}
 	}
 	
 	init(training: TrainingModel) {
 		self.training = training
-		self._viewModel = State(initialValue: ViewModel(training: training))
 	}
 	
 	private func scrollToBottom(_ proxy: ScrollViewProxy) {
@@ -85,77 +84,5 @@ struct TrainingView: View {
 			.modelContainer(container)
 	} catch {
 		fatalError("Failed to create model container")
-	}
-}
-
-struct ExerciseView: View {
-	
-	@Bindable var exercise: ExerciseModel
-	
-	@State var viewModel: TrainingView.ViewModel
-	@State private var isExpanded = true
-	
-	let showTextField: Bool
-	let completion: () -> Void
-	
-	var body: some View {
-		VStack {
-			HStack {
-				Text(exercise.name)
-					.frame(maxWidth: .infinity, alignment: .leading)
-					.font(.title2)
-				
-				Text("Max: ") +
-				Text(ExerciseRecordManager.shared.getBestForExerciseWith(uuid: exercise.uuid), format: .number) +
-				Text(" \(exercise.mainStat.unit)")
-				
-				Spacer(minLength: 0)
-				
-				if !showTextField {
-					Image(systemName: "chevron.up")
-						.rotationEffect(.init(degrees: isExpanded ? 0 : 180))
-						.foregroundStyle(.indigo)
-				}
-			}
-			.contentShape(.rect)
-			.onTapGesture {
-				if !showTextField {
-					withAnimation(.bouncy) {
-						isExpanded.toggle()
-					}
-				}
-			}
-			
-			if isExpanded {
-				Rectangle()
-					.frame(height: 1)
-					.foregroundStyle(.secondary)
-				
-				ForEach(Array(zip(exercise.sets.indices, exercise.sets)), id: \.0) { index, item in
-					HStack {
-						Text("\(index + 1).")
-						
-						Text("\(item.description)")
-							.frame(maxWidth: .infinity, alignment: .leading)
-					}
-					
-					Rectangle()
-						.frame(height: 1)
-						.foregroundStyle(.secondary)
-				}
-			}
-			
-			if showTextField {
-				SetEditorView(exercise: exercise, completion: completion)
-			}
-		}
-		.padding()
-		.background(HierarchicalShapeStyle.quaternary)
-		.overlay {
-			RoundedRectangle(cornerRadius: 15, style: .continuous)
-				.stroke(.indigo, lineWidth: 5)
-		}
-		.clipShape(.rect(cornerRadius: 15, style: .continuous))
-		.padding(.horizontal)
 	}
 }
